backend/requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==3.2.17
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.0.4
idna==3.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.4.0
psycopg2-binary==2.8.6
pycparser==2.21
pydantic==1.10.5
pyflakes==3.0.1
PyJWT==2.1.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2020.1
regex==2022.3.2
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.4.0
unify==0.5
untokenize==0.1.1
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.12.1
